+++ tests/style/bug_131/Potato.m
@@ -1,0 +1,5 @@
+% (c) Copyright 2020 Florian Schanda
+classdef Potato
+    enumeration Picasso, Laura, King_Edward
+    end
+end